Separated LinkedVector and its related entities into linked_vector.rs
