+++ docs/doc/implementors/core/marker/trait.Freeze.js
@@ -1,3 +1,3 @@
 (function() {var implementors = {
-"linked_vector":[["impl&lt;'a, T&gt; Freeze for <a class=\"struct\" href=\"linked_vector/struct.Cursor.html\" title=\"struct linked_vector::Cursor\">Cursor</a>&lt;'a, T&gt;",1,["linked_vector::cursor::Cursor"]],["impl&lt;'a, T&gt; Freeze for <a class=\"struct\" href=\"linked_vector/struct.CursorMut.html\" title=\"struct linked_vector::CursorMut\">CursorMut</a>&lt;'a, T&gt;",1,["linked_vector::cursor::CursorMut"]],["impl Freeze for <a class=\"struct\" href=\"linked_vector/struct.HNode.html\" title=\"struct linked_vector::HNode\">HNode</a>",1,["linked_vector::HNode"]],["impl&lt;T&gt; Freeze for <a class=\"struct\" href=\"linked_vector/struct.LinkedVector.html\" title=\"struct linked_vector::LinkedVector\">LinkedVector</a>&lt;T&gt;",1,["linked_vector::LinkedVector"]],["impl&lt;'a, T&gt; Freeze for <a class=\"struct\" href=\"linked_vector/struct.Handles.html\" title=\"struct linked_vector::Handles\">Handles</a>&lt;'a, T&gt;",1,["linked_vector::Handles"]],["impl&lt;'a, T&gt; Freeze for <a class=\"struct\" href=\"linked_vector/struct.Iter.html\" title=\"struct linked_vector::Iter\">Iter</a>&lt;'a, T&gt;",1,["linked_vector::Iter"]],["impl&lt;'a, T&gt; Freeze for <a class=\"struct\" href=\"linked_vector/struct.IterMut.html\" title=\"struct linked_vector::IterMut\">IterMut</a>&lt;'a, T&gt;",1,["linked_vector::IterMut"]],["impl&lt;T&gt; Freeze for <a class=\"struct\" href=\"linked_vector/struct.IntoIter.html\" title=\"struct linked_vector::IntoIter\">IntoIter</a>&lt;T&gt;",1,["linked_vector::IntoIter"]]]
+"linked_vector":[["impl&lt;'a, T&gt; Freeze for <a class=\"struct\" href=\"linked_vector/struct.Cursor.html\" title=\"struct linked_vector::Cursor\">Cursor</a>&lt;'a, T&gt;",1,["linked_vector::cursor::Cursor"]],["impl&lt;'a, T&gt; Freeze for <a class=\"struct\" href=\"linked_vector/struct.CursorMut.html\" title=\"struct linked_vector::CursorMut\">CursorMut</a>&lt;'a, T&gt;",1,["linked_vector::cursor::CursorMut"]],["impl Freeze for <a class=\"struct\" href=\"linked_vector/struct.HNode.html\" title=\"struct linked_vector::HNode\">HNode</a>",1,["linked_vector::linked_vector::HNode"]],["impl&lt;T&gt; Freeze for <a class=\"struct\" href=\"linked_vector/struct.LinkedVector.html\" title=\"struct linked_vector::LinkedVector\">LinkedVector</a>&lt;T&gt;",1,["linked_vector::linked_vector::LinkedVector"]],["impl&lt;'a, T&gt; Freeze for <a class=\"struct\" href=\"linked_vector/struct.Handles.html\" title=\"struct linked_vector::Handles\">Handles</a>&lt;'a, T&gt;",1,["linked_vector::linked_vector::Handles"]],["impl&lt;'a, T&gt; Freeze for <a class=\"struct\" href=\"linked_vector/struct.Iter.html\" title=\"struct linked_vector::Iter\">Iter</a>&lt;'a, T&gt;",1,["linked_vector::linked_vector::Iter"]],["impl&lt;'a, T&gt; Freeze for <a class=\"struct\" href=\"linked_vector/struct.IterMut.html\" title=\"struct linked_vector::IterMut\">IterMut</a>&lt;'a, T&gt;",1,["linked_vector::linked_vector::IterMut"]],["impl&lt;T&gt; Freeze for <a class=\"struct\" href=\"linked_vector/struct.IntoIter.html\" title=\"struct linked_vector::IntoIter\">IntoIter</a>&lt;T&gt;",1,["linked_vector::linked_vector::IntoIter"]]]
 };if (window.register_implementors) {window.register_implementors(implementors);} else {window.pending_implementors = implementors;}})()
