+++ docs/doc/linked_vector/linked_vector/struct.IntoIter.html
@@ -1,0 +1,11 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta http-equiv="refresh" content="0;URL=../../linked_vector/struct.IntoIter.html">
+    <title>Redirection</title>
+</head>
+<body>
+    <p>Redirecting to <a href="../../linked_vector/struct.IntoIter.html">../../linked_vector/struct.IntoIter.html</a>...</p>
+    <script>location.replace("../../linked_vector/struct.IntoIter.html" + location.search + location.hash);</script>
+</body>
+</html>
