+++ docs/doc/linked_vector/linked_vector/struct.LinkedVector.html
@@ -1,0 +1,11 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta http-equiv="refresh" content="0;URL=../../linked_vector/struct.LinkedVector.html">
+    <title>Redirection</title>
+</head>
+<body>
+    <p>Redirecting to <a href="../../linked_vector/struct.LinkedVector.html">../../linked_vector/struct.LinkedVector.html</a>...</p>
+    <script>location.replace("../../linked_vector/struct.LinkedVector.html" + location.search + location.hash);</script>
+</body>
+</html>
