+++ docs/doc/source-files.js
@@ -1,4 +1,4 @@
 var sourcesIndex = JSON.parse('{\
-"linked_vector":["",[],["cursor.rs","lib.rs"]]\
+"linked_vector":["",[],["cursor.rs","lib.rs","linked_vector.rs"]]\
 }');
 createSourceSidebar();
