+++ docs/doc/src/linked_vector/cursor.rs.html
@@ -302,8 +302,12 @@
 <a href="#302" id="302">302</a>
 <a href="#303" id="303">303</a>
 <a href="#304" id="304">304</a>
+<a href="#305" id="305">305</a>
+<a href="#306" id="306">306</a>
+<a href="#307" id="307">307</a>
+<a href="#308" id="308">308</a>
 </pre><pre class="rust"><code>
-<span class="kw">use crate</span>::<span class="kw-2">*</span>;
+<span class="kw">use </span><span class="kw">crate</span>::linked_vector::<span class="kw-2">*</span>;
 
 <span class="doccomment">/// A cursor is a position within a linked vector. It can be used to traverse
 /// the list in either direction, and to access the element at the current
@@ -358,13 +362,17 @@
 
     <span class="doccomment">/// Moves the cursor forward by the specified number of elements. Returns
     /// the handle of the element at the new position if the cursor was moved,
-    /// Err(handle) if the cursor was already at the end of the list.
+    /// Err(handle) if the cursor did not move forward by the specified amount.
+    /// The handle at the current position after the move is returned in 
+    /// either Result variant.
     /// 
     </span><span class="kw">fn </span>forward(<span class="kw-2">&amp;mut </span><span class="self">self</span>, n: usize) -&gt; <span class="prelude-ty">Result</span>&lt;HNode, HNode&gt;;
 
     <span class="doccomment">/// Moves the cursor backward by the specified number of elements. Returns
     /// the handle of the element at the new position if the cursor was moved,
-    /// Err(handle) if the cursor was already at the start of the list.
+    /// Err(handle) if the cursor did not move backward by the specified amount.
+    /// The handle at the current position after the move is returned in 
+    /// either Result variant.
     /// 
     </span><span class="kw">fn </span>backward(<span class="kw-2">&amp;mut </span><span class="self">self</span>, n: usize) -&gt; <span class="prelude-ty">Result</span>&lt;HNode, HNode&gt;;
 }
